PubMatic RTD Provider: create Unified Pricing Rules plugin
