+++ libraries/pubmaticUtils/plugins/bidderOptimization.js
@@ -4,16 +4,16 @@ import { getBrowserType } from '../pubmaticUtils.js';
 import { logInfo, logError } from '../../../src/utils.js';
 
 let config = null;
-let CONSTANTS = null;
+const CONSTANTS = Object.freeze({
+  LOG_PRE_FIX: 'PubMatic-Bidder-Optimization: '
+});
 
 /**
  * Initialize the bidder optimization plugin
  * @param {Object} bidderOptimisationConfig - Bidder optimization configuration
- * @param {Object} constants - Constants object
  * @returns {Promise<boolean>} - Promise resolving to initialization status
  */
-export async function init(bidderOptimisationConfig, constants) {
-  CONSTANTS = constants;
+export async function init(bidderOptimisationConfig) {
   
   // Process bidder optimization configuration
   try {
@@ -54,40 +54,42 @@ export async function processBidRequest(reqBidsConfigObj) {
   }
 }
 
-
 /**
  * Get targeting data
  * @param {Array} adUnitCodes - Ad unit codes
+ * @param {Object} config - Module configuration
+ * @param {Object} userConsent - User consent data
  * @param {Object} auction - Auction object
  * @returns {Object} - Targeting data
  */
-export function getTargeting(adUnitCodes, auction) {
+export function getTargeting(adUnitCodes, config, userConsent, auction) {
   // Implementation for targeting data, if not applied then do nothing
-}
-
-/**
- * Filter bidders from ad units
- * @param {Array} bidderList - List of bidders to filter
- * @param {Object} reqBidsConfigObj - Bid request config object
- * @param {string} adUnitCode - Ad unit code
- */
-export function filterBidders(bidderList, reqBidsConfigObj, adUnitCode) {
-  if (!reqBidsConfigObj.adUnits || !Array.isArray(reqBidsConfigObj.adUnits)) {
-    return;
-  }
-
-  const adUnit = reqBidsConfigObj.adUnits.find(unit => unit.code === adUnitCode);
-  if (!adUnit || !adUnit.bids || !Array.isArray(adUnit.bids)) {
-    return;
-  }
-
-  // Filter out specified bidders
-  adUnit.bids = adUnit.bids.filter(bid => !bidderList.includes(bid.bidder));
 }
 
 // Export the bidder optimization functions
 export const BidderOptimization = {
   init,
   processBidRequest,
-  filterBidders
+  getTargeting
+};
+
+/**
+ * Filter out specified bidders from adUnits with matching code
+ * @param {Array} bidderList - List of bidder names to be filtered out
+ * @param {Object} reqBidsConfigObj - The bid request configuration object
+ * @param {string} adUnitCode - The code of the adUnit to filter bidders from
+ */
+export const filterBidders = (bidderList, reqBidsConfigObj, adUnitCode) => {
+  // Validate inputs
+  if (!bidderList || !Array.isArray(bidderList) || bidderList.length === 0 ||
+      !reqBidsConfigObj || !reqBidsConfigObj.adUnits || !Array.isArray(reqBidsConfigObj.adUnits)) {
+    return;
+  }
+  // Find the adUnit with the matching code
+  const adUnit = reqBidsConfigObj.adUnits.find(unit => unit.code === adUnitCode);
+
+  // If adUnit exists and has bids array, filter out the specified bidders
+  if (adUnit && adUnit.bids && Array.isArray(adUnit.bids)) {
+    adUnit.bids = adUnit.bids.filter(bid => !bidderList.includes(bid.bidder));
+  }
 };
